backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pyodbc==5.2.0
python-multipart==0.0.12
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
